limepeza do codigo, troca de png para webp, retirada de codigos desnecessarios e simplificação de js

+++ assets/js/menu.js
@@ -1,5 +1,8 @@
-const botaoReset = document.querySelector(".reset")
-
-botaoReset.addEventListener("click", ()=>{
-    localStorage.clear()
-})
+const botaoReset = document.querySelector(".reset");
+botaoReset.addEventListener("click", () => {
+    if (confirm("Tem certeza de que deseja resetar o progresso? Isso limpará todos os dados salvos.")) {
+        localStorage.clear();
+        alert("Progresso resetado com sucesso.");
+        location.reload();
+    }
+});
